requirements.txt
pystray
Pillow
psutil
requests
pyinstaller
